requirements.txt
streamlit>=1.32
pandas>=2.2
openpyxl>=3.1
numpy>=1.26
reportlab>=4.0
matplotlib>=3.8
plotly==5.24.1
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
kaleido==0.2.1